requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# Graph Database
neo4j>=5.0.0
neo4j-graphrag>=1.0.0

# Data Processing
duckdb>=0.9.0
pandas>=2.0.0
openpyxl>=3.1.0

# Document Parsing
pypdf>=3.17.0

# LLM & Orchestration
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.6.0

# Vector Embeddings
sentence-transformers>=2.2.0

# Type Checking & Quality
mypy>=1.7.0
types-requests>=2.31.0

# Utilities
pydantic>=2.0.0
tqdm>=4.66.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
flake8>=6.1.0
black>=23.12.0
